Factorize and rename Output java8 streaming feature to a single class StreamDecoder
