+++ java8/src/test/java/feign/stream/StreamDecoderTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package feign.optionals;
+package feign.stream;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -28,7 +28,6 @@ import feign.Feign;
 import feign.RequestLine;
 import feign.codec.DecodeException;
 import feign.jackson.JacksonIterator;
-import feign.stream.Java8StreamDecoder;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 
@@ -67,7 +66,7 @@ public class StreamDecoderTest {
     server.enqueue(new MockResponse().setBody("foo\nbar"));
 
     StreamInterface api = Feign.builder()
-        .decoder(new Java8StreamDecoder((type, response) -> {
+        .decoder(new StreamDecoder((type, response) -> {
           try {
             return new BufferedReader(new InputStreamReader(response.body().asInputStream())).lines().iterator();
           } catch (IOException e) {
@@ -88,7 +87,7 @@ public class StreamDecoderTest {
     ObjectMapper mapper = new ObjectMapper();
 
     StreamInterface api = Feign.builder()
-        .decoder(new Java8StreamDecoder((type, response) -> JacksonIterator.<StreamInterface.Car>builder().of(type).mapper(mapper).response(response).build()))
+        .decoder(new StreamDecoder((type, response) -> JacksonIterator.<StreamInterface.Car>builder().of(type).mapper(mapper).response(response).build()))
               .target(StreamInterface.class, server.url("/").toString());
 
     try (Stream<StreamInterface.Car> stream = api.getCars()) {
